# api/openapi/platform-v1.yaml
components:
  schemas:
...
    ResourceGraphSuccessV2:
      type: object
      additionalProperties: false
      properties:
        requestId:
          $ref: '#/components/schemas/Identifier'
        data:
          $ref: '#/components/schemas/ResourceGraphV2'
        meta:
          type: object
          additionalProperties: true
      required:
      - requestId
      - data
...
    EvidenceSuccessV2:
      type: object
      additionalProperties: false
      properties:
        requestId:
          $ref: '#/components/schemas/Identifier'
        data:
          $ref: '#/components/schemas/EvidenceV2'
        meta:
          type: object
          additionalProperties: true
      required:
      - requestId
      - data
...
    DiagnosticGraphBuildRequest:
      additionalProperties: false
      properties:
        entryCanonicalId:
          $ref: '#/components/schemas/CanonicalID'
        policy:
          $ref: '#/components/schemas/GraphBudget'
        recipe:
          enum:
          - incident
          type: string
      required:
      - entryCanonicalId
      - recipe
      - policy
      type: object
    ErrorEnvelope:
      additionalProperties: false
      properties:
        code:
          enum:
          - INVALID_ARGUMENT
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - PRECONDITION_FAILED
          - RATE_LIMITED
          - SOURCE_DEGRADED
          - GRAPH_NOT_READY
          - GRAPH_PARTIAL
          - TIMEOUT
          - BUDGET_EXHAUSTED
          - POLICY_DENIED
          - RISK_ACK_REQUIRED
          - RISK_ACK_COMMAND_MISMATCH
          - STEP_UP_REQUIRED
          - STEP_UP_EXPIRED
          - IDEMPOTENCY_CONFLICT
          - INTERNAL
          - CAPABILITY_DISABLED
          - SOURCE_NOT_CONFIGURED
          - IDEMPOTENCY_IN_PROGRESS
          - IDEMPOTENCY_KEY_REUSED
          - EXECUTION_PROFILE_DENIED
          - COMMAND_POLICY_DENIED
          - EXECUTION_UNKNOWN
          - OUTPUT_TRUNCATED
          - BUNDLE_SIGNATURE_INVALID
          - BUNDLE_DIGEST_MISMATCH
          - BUNDLE_EXTRACTION_REJECTED
          - PROFILE_COMPONENT_CONFLICT
          - UPSTREAM_COMPATIBILITY_UNVERIFIED
          - OPENBAO_SEALED
          - HOST_NOT_ONBOARDED
          type: string
        details:
          additionalProperties: true
          type: object
        message:
          type: string
        requestId:
          $ref: '#/components/schemas/Identifier'
        retryable:
          type: boolean
      required:
      - code
      - message
      - requestId
      - retryable
      type: object
...
    SP04ErrorEnvelopeV2:
      additionalProperties: false
      properties:
        code:
          enum:
          - INVALID_ARGUMENT
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - PRECONDITION_FAILED
          - RATE_LIMITED
          - SOURCE_DEGRADED
          - GRAPH_NOT_READY
          - GRAPH_PARTIAL
          - TIMEOUT
          - BUDGET_EXHAUSTED
          - POLICY_DENIED
          - RISK_ACK_REQUIRED
          - RISK_ACK_COMMAND_MISMATCH
          - STEP_UP_REQUIRED
          - STEP_UP_EXPIRED
          - IDEMPOTENCY_CONFLICT
          - INTERNAL
          - CAPABILITY_DISABLED
          - SOURCE_NOT_CONFIGURED
          - IDEMPOTENCY_IN_PROGRESS
          - IDEMPOTENCY_KEY_REUSED
          - EXECUTION_PROFILE_DENIED
          - COMMAND_POLICY_DENIED
          - EXECUTION_UNKNOWN
          - OUTPUT_TRUNCATED
          - BUNDLE_SIGNATURE_INVALID
          - BUNDLE_DIGEST_MISMATCH
          - BUNDLE_EXTRACTION_REJECTED
          - PROFILE_COMPONENT_CONFLICT
          - UPSTREAM_COMPATIBILITY_UNVERIFIED
          - OPENBAO_SEALED
          - HOST_NOT_ONBOARDED
          - SOURCE_SCOPE_UNVERIFIED
          - STALE_CONTEXT
          - CURSOR_EXPIRED
          type: string
        details:
          additionalProperties: true
          type: object
        message:
          type: string
        requestId:
          $ref: '#/components/schemas/Identifier'
        retryable:
          type: boolean
        schemaVersion:
          type: string
          enum:
          - error-envelope/v2
      required:
      - code
      - message
      - requestId
      - retryable
      - schemaVersion
      type: object
    LegalHoldRequestV1:
      type: object
      additionalProperties: false
      required:
      - evidenceId
      - hold
      properties:
        evidenceId:
          type: string
          format: uuid
        hold:
          type: boolean
...
    LegalHoldPageV1:
      type: object
      additionalProperties: false
      required:
      - schemaVersion
      - requestId
      - data
      - meta
      properties:
        schemaVersion:
          type: string
          enum:
          - legal-hold-page/v1
        requestId:
          type: string
          minLength: 1
        data:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
            - evidenceId
            - hold
            properties:
              evidenceId:
                type: string
                format: uuid
              hold:
                type: boolean
        meta:
          type: object
          additionalProperties: false
          required:
          - nextCursor
          - freshness
          - partial
          - warnings
          - degradedSources
          properties:
            nextCursor:
              type: string
            freshness:
              type: string
              enum:
              - fresh
              - stale
              - unavailable
            partial:
              type: boolean
            warnings:
              type: array
              items:
                type: string
            degradedSources:
              type: array
              items:
                type: string
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: Smart Operations Platform API
  version: 1.3.0
...
openapi: 3.1.0
paths:
...
  /api/v1/admin/legal-holds:
    get:
      operationId: getLegalHolds
      parameters:
      - description: Opaque cursor for keyset pagination.
        in: query
        name: cursor
        schema:
          type: string
      - in: query
        name: limit
        schema:
          minimum: 1
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalHoldPageV1'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getLegalHolds
      tags:
      - Admin
    post:
      operationId: createLegalHold
      parameters:
      - description: Stable key for replay-safe write handling.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegalHoldRequestV1'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalHoldResultV1'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: createLegalHold
      tags:
      - Admin
      x-idempotency: required
...
  /api/v1/diagnostic-graphs:build:
    post:
      operationId: buildDiagnosticGraph
      parameters:
      - description: Stable key for replay-safe write handling.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosticGraphBuildRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGraphSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: buildDiagnosticGraph
      tags:
      - DiagnosticGraphs
      x-idempotency: required
  /api/v1/evidence/{evidenceId}:
    get:
      operationId: getEvidence
      parameters:
      - in: path
        name: evidenceId
        required: true
        schema:
          format: identifier
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getEvidence
      tags:
      - Evidence
  /api/v1/evidence:query:
    post:
      operationId: queryEvidence
      parameters:
      - description: Stable key for replay-safe write handling.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceQueryRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResultSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: queryEvidence
      tags:
      - Evidence
      x-idempotency: required
...
  /api/v1/resources:
    get:
      operationId: getResources
      parameters:
      - in: query
        name: kind
        schema:
          type: string
      - in: query
        name: clusterUid
        schema:
          type: string
      - in: query
        name: namespace
        schema:
          type: string
      - in: query
        name: name
        schema:
          type: string
      - in: query
        name: health
        schema:
          type: string
      - in: query
        name: label
        schema:
          type: string
      - description: Opaque cursor for keyset pagination.
        in: query
        name: cursor
        schema:
          type: string
      - in: query
        name: limit
        schema:
          minimum: 1
          type: integer
      - in: query
        name: sort
        schema:
          type: string
      - in: query
        name: order
        schema:
          enum:
          - asc
          - desc
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePageV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getResources
      tags:
      - Resources
  /api/v1/resources/by-canonical-id:
    get:
      operationId: getResourceByCanonicalID
      parameters:
      - allowReserved: false
        description: Percent-encode the complete Canonical ID value per RFC 3986; the server decodes and validates it.
        in: query
        name: canonicalId
        required: true
        schema:
          $ref: '#/components/schemas/CanonicalID'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGraphSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getResourceByCanonicalID
      tags:
      - Resources
  /api/v1/resources/impact-scope:
    get:
      operationId: getResourceImpactScope
      parameters:
      - allowReserved: false
        description: Percent-encode the complete Canonical ID value per RFC 3986; the server decodes and validates it.
        in: query
        name: canonicalId
        required: true
        schema:
          $ref: '#/components/schemas/CanonicalID'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGraphSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getResourceImpactScope
      tags:
      - Resources
  /api/v1/resources/neighbors:
    get:
      operationId: getResourceNeighbors
      parameters:
      - allowReserved: false
        description: Percent-encode the complete Canonical ID value per RFC 3986; the server decodes and validates it.
        in: query
        name: canonicalId
        required: true
        schema:
          $ref: '#/components/schemas/CanonicalID'
      - in: query
        name: direction
        schema:
          enum:
          - in
          - out
          - both
          type: string
      - in: query
        name: relationKind
        schema:
          type: string
      - in: query
        name: depth
        schema:
          minimum: 0
          type: integer
      - in: query
        name: maxNodes
        schema:
          minimum: 1
          type: integer
      - in: query
        name: maxEdges
        schema:
          minimum: 1
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGraphSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getResourceNeighbors
      tags:
      - Resources
security:
- BearerAuth: []
servers:
- url: /